refactor(appstore): rename to app hub (#4917)
