+++ dhis-2/dhis-services/dhis-service-administration/src/main/java/org/hisp/dhis/apphub/AppVersion.java
@@ -1,4 +1,4 @@
-package org.hisp.dhis.appstore;
+package org.hisp.dhis.apphub;
 
 /*
  * Copyright (c) 2004-2020, University of Oslo
@@ -57,6 +57,7 @@ public class AppVersion
 
     public AppVersion()
     {
+        // empty constructor
     }
 
     @JsonIgnore
